Move githubconnector to shared. Think I found my mistake.
