+++ shared/githubconnector/get_releases.go
@@ -1,4 +1,4 @@
-package github_connector
+package githubconnector
 
 import (
 	"context"
